+++ rumqttc-v4/src/eventloop.rs
@@ -42,7 +42,7 @@ use {
 ))]
 use crate::websockets::split_url_with_default_port;
 
-#[cfg(feature = "proxy")]
+#[cfg(any(feature = "http-proxy", feature = "socks-proxy"))]
 use crate::proxy::ProxyError;
 
 #[derive(Debug)]
@@ -208,7 +208,7 @@ pub enum ConnectionError {
     #[cfg(feature = "websocket")]
     #[error("Invalid Url: {0}")]
     InvalidUrl(#[from] UrlError),
-    #[cfg(feature = "proxy")]
+    #[cfg(any(feature = "http-proxy", feature = "socks-proxy"))]
     #[error("Proxy Connect: {0}")]
     Proxy(#[from] ProxyError),
     #[cfg(feature = "websocket")]
@@ -1700,7 +1700,7 @@ async fn network_connect(
     };
 
     let tcp_stream: Box<dyn AsyncReadWrite> = {
-        #[cfg(feature = "proxy")]
+        #[cfg(any(feature = "http-proxy", feature = "socks-proxy"))]
         if let Some(proxy) = options.proxy() {
             proxy
                 .connect(
@@ -1714,7 +1714,7 @@ async fn network_connect(
             let addr = format!("{domain}:{port}");
             options.socket_connect(addr, network_options).await?
         }
-        #[cfg(not(feature = "proxy"))]
+        #[cfg(not(any(feature = "http-proxy", feature = "socks-proxy")))]
         {
             let addr = format!("{domain}:{port}");
             options.socket_connect(addr, network_options).await?